pyproject.toml
[project]
name = "test-saml-idp"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Paul Watts", email = "paulcwatts@gmail.com"},
]
dependencies = [
    "fastapi[standard]>=0.112.1",
    "jinja2>=3.1.4",
    "lxml>=5.3.0",
    "pydantic-settings>=2.4.0",
    "signxml>=4.0.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "fastapi dev"
run = "fastapi run"
test = "pytest"
lint-check = "ruff check ."
type-check = "pyright"
format-check = "black --check ."
lint.composite = ["lint-check", "type-check", "format-check"]
lint.keep_going = true
fix-all.composite = [
    "ruff check --fix .",
    "black ."
]

[tool.pdm.dev-dependencies]
test = [
    "black>=24.8.0",
    "pyright>=1.1.377",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.6.2",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN401", "D203", "D211", "D212", "D213", "S320", "PLR0913", "TRY300"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D100", "D104", "S"]

[tool.pyright]
pythonVersion = "3.12"
disableBytesTypePromotions = true
strictListInference = true
strictDictionaryInference = true
structSetInference = true
reportMissingTypeArgument = true
reportMatchNotExhaustive = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportUnnecessaryIsInstance = true
reportUnnecessaryTypeIgnoreComment = true
# signxml and lxml have too many import problems
reportPrivateImportUsage = false
reportAttributeAccessIssue = false

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "session"
